Move common email/text methods into new common base renewal reminders class
