+++ app/services/renewal_reminders/renewal_reminder_email_service_base.rb
@@ -2,7 +2,7 @@
 
 module RenewalReminders
 
-  class RenewalReminderEmailServiceBase < WasteExemptionsEngine::BaseService
+  class RenewalReminderEmailServiceBase < RenewalReminderServiceBase
     def run
       expiring_registrations.each do |registration|
         send_email(registration)
@@ -22,21 +22,6 @@ module RenewalReminders
       default_scope.where(
         id: all_active_exemptions_registration_ids
       ).contact_email_present.site_address_is_not_nccc
-    end
-
-    def all_active_exemptions_registration_ids
-      WasteExemptionsEngine::RegistrationExemption
-        .all_active_exemptions
-        .where(expires_on: expires_in_days.days.from_now.to_date)
-        .pluck(:registration_id)
-    end
-
-    def default_scope
-      WasteExemptionsEngine::Registration
-    end
-
-    def expires_in_days
-      raise(NotImplementedError)
     end
   end
 end
